# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ocean Blog API",
        "description": "Blogs API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Dev version"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Отримати список категорій",
                "description": "Повертає всі категорії у форматі JSON.",
                "operationId": "ba2f88513b5f6d9f04554acafeaaa2ac",
                "responses": {
                    "200": {
                        "description": "Успішно отримано список категорій",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Створити нову категорію",
                "description": "Створює нову категорію та повертає її у форматі JSON.",
                "operationId": "93f351bf6772f61fa80b47d8b7b531f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Категорію створено успішно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Показати категорію",
                "description": "Повертає одну категорію за ID.",
                "operationId": "554c1c4fb0cbf318d3079ba7e84d38dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категорії",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішно отримано категорію",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Категорію не знайдено"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Оновити категорію",
                "description": "Оновлює існуючу категорію за ID.",
                "operationId": "3cbc0d5fc107a29ea3a9b67c220746c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категорії",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Категорію оновлено успішно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Категорію не знайдено"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Видалити категорію",
                "description": "Видаляє категорію за ID.",
                "operationId": "6b203a9d1a8c8aa75716dc24525d8c84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категорії",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Категорію видалено успішно"
                    },
                    "404": {
                        "description": "Категорію не знайдено"
                    }
                }
            }
        },
        "/posts/{post}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get list of comments",
                "description": "Returns paginated list of comments for a post.",
                "operationId": "07ed68f9df3dcc788051290f7b59608a",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of comments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comment"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{post}/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Show comment",
                "description": "Returns a single comment by ID for a post.",
                "operationId": "b965e1085cdff96b7ae18512600e83ae",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved comment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            }
        },
        "/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment",
                "description": "Creates a new comment for a post and returns it as JSON.",
                "operationId": "918217a8d2f7d294c0778a1d08bc7033",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update comment",
                "description": "Updates an existing comment by ID for a post.",
                "operationId": "bc7846bea2d91107da67e2133958fe3a",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete comment",
                "description": "Deletes a comment by ID for a post.",
                "operationId": "eba10559c9b6a621e4f5764a8b8a4a10",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment deleted successfully"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "News"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-03T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-03T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "Comment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "This is a comment."
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-03T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-03T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "CategoryCreateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "News"
                    },
                    "description": {
                        "type": "string",
                        "example": "Latest news and updates"
                    },
                    "slug": {
                        "type": "string",
                        "example": "news"
                    }
                },
                "type": "object"
            },
            "CategoryUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "News"
                    },
                    "description": {
                        "type": "string",
                        "example": "Latest news and updates"
                    },
                    "slug": {
                        "type": "string",
                        "example": "news"
                    }
                },
                "type": "object"
            },
...
            "CommentCreateRequest": {
                "properties": {
                    "content": {
                        "type": "string",
                        "example": "This is a comment."
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CommentUpdateRequest": {
                "properties": {
                    "content": {
                        "type": "string",
                        "example": "This is an updated comment."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}